BOB-35: Renaming GoogleCalendarClient to GoogleClient
