+++ index.js
@@ -1,7 +1,7 @@
 const AtlassianClientFactory = require('./src/clients/AtlassianClientFactory');
 const CircleciClientFactory = require('./src/clients/CircleciClientFactory');
 const GithubClientFactory = require('./src/clients/GithubClientFactory');
-const GoogleCalendarClientFactory = require('./src/clients/GoogleCalendarClientFactory');
+const GoogleClientFactory = require('./src/clients/GoogleClientFactory');
 const RedisClientFactory = require('./src/clients/RedisClientFactory');
 const SlackClientFactory = require('./src/clients/SlackClientFactory');
 const JiraClient = require('./src/clients/JiraClient');
@@ -44,7 +44,7 @@ module.exports = {
     AtlassianClientFactory,
     CircleciClientFactory,
     GithubClientFactory,
-    GoogleCalendarClientFactory,
+    GoogleClientFactory,
     RedisClientFactory,
     SlackClientFactory,
     JiraClient,
